server/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
mysqlclient
pandas
beautifulsoup4
certifi==2022.12.7
charset-normalizer==3.0.1
et-xmlfile==1.1.0
idna==3.4
numpy
openpyxl==3.1.1
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
soupsieve==2.4
urllib3==1.26.14
requests
openpyxl